finished-sprint-11

+++ src/utils/Api.js
@@ -10,18 +10,19 @@ class Api {
     if (res.ok) {
       return res.json();
     } else {
-      return Promise.reject(`Something is wrong: 4 8 15 16 23 42 && ${res.status} ${res.statusText}`);
+      return Promise.reject(
+        `Something is wrong: 4 8 15 16 23 42 && ${res.status} ${res.statusText}`
+      );
     }
   }
   /**
-   * getting profile info from the server
+   * @description getting profile info from the server
    */
   getProfileInfo() {
     return fetch(`${this._url}/users/me`, {
       method: 'GET',
-      headers: this._headers
-    })
-      .then(this._checkErrors);
+      headers: this._headers,
+    }).then(this._checkErrors);
   }
   /**
    * getting cards from the server
@@ -29,9 +30,8 @@ class Api {
   getCards() {
     return fetch(`${this._url}/cards`, {
       method: 'GET',
-      headers: this._headers
-    })
-    .then(this._checkErrors);
+      headers: this._headers,
+    }).then(this._checkErrors);
   }
   /**
    * editing user profile info on the server
@@ -41,11 +41,10 @@ class Api {
       method: 'PATCH',
       headers: this._headers,
       body: JSON.stringify({
-        name: info[`profile-name`],
-        about: info[`profile-job`]
-      })
-    })
-      .then(this._checkErrors);
+        name: info.name,
+        about: info.about,
+      }),
+    }).then(this._checkErrors);
   }
   /**
    * updating profile avatar on the server
@@ -55,10 +54,9 @@ class Api {
       method: 'PATCH',
       headers: this._headers,
       body: JSON.stringify({
-        avatar: avatarUrl
-      })
-    })
-      .then(this._checkErrors);
+        avatar: avatarUrl,
+      }),
+    }).then(this._checkErrors);
   }
   /**
    * adding a new card to the server
@@ -69,10 +67,9 @@ class Api {
       headers: this._headers,
       body: JSON.stringify({
         name: name,
-        link: link
-      })
-    })
-    .then(this._checkErrors);
+        link: link,
+      }),
+    }).then(this._checkErrors);
   }
   /**
    * removing a card from the server
@@ -81,28 +78,25 @@ class Api {
     return fetch(`${this._url}/cards/${cardId}`, {
       method: 'DELETE',
       headers: this._headers,
-    })
-    .then(this._checkErrors);
+    }).then(this._checkErrors);
   }
   /**
-   * adding like to the server
+   * @description this method sends a fetch request adding or removing like depending on whether a card is liked or not
+   * @param {*} cardId is used to identify a card that is to changed
+   * @param {*} isLiked is used to identify whether a card is liked or not
    */
-  addLike(cardId) {
-    return fetch(`${this._url}/cards/likes/${cardId}`, {
-      method: 'PUT',
-      headers: this._headers,
-    })
-      .then(this._checkErrors);
-  }
-  /**
-   * removing like from the server
-   */
-  removeLike(cardId) {
-    return fetch(`${this._url}/cards/likes/${cardId}`, {
-      method: 'DELETE',
-      headers: this._headers,
-    })
-      .then(this._checkErrors);
+  changeLikeCardStatus(cardId, isLiked) {
+    if (isLiked) {
+      return fetch(`${this._url}/cards/likes/${cardId}`, {
+        method: 'DELETE',
+        headers: this._headers,
+      }).then(this._checkErrors);
+    } else {
+      return fetch(`${this._url}/cards/likes/${cardId}`, {
+        method: 'PUT',
+        headers: this._headers,
+      }).then(this._checkErrors);
+    }
   }
 }
 
@@ -110,7 +104,7 @@ const api = new Api({
   url: 'https://mesto.nomoreparties.co/v1/cohort-15',
   headers: {
     authorization: '1ed91742-56fd-4a56-812b-580db32d6be2',
-    'Content-Type': 'application/json'
-  }
+    'Content-Type': 'application/json',
+  },
 });
 export default api;
